refactor(frontend): extract setup wizard load and action helpers from use-setup
512-line hook becomes 309 lines over setup-load.ts and
setup-actions.ts (verbatim moves; hook order unchanged). Completed
after the extraction agent was cut off mid-verification; all gates
re-run green.
