+++ frontend/src/features/setup/use-setup.ts
@@ -1,10 +1,9 @@
 "use client";
 
-import { Effect, Result } from "effect";
+import { Effect } from "effect";
 
 import { useCallback, useMemo, useState } from "react";
 import { useRouter } from "next/navigation";
-import api from "@/lib/api/client";
 import type {
   EngineBackend,
   EngineJob,
@@ -16,26 +15,23 @@ import type {
 } from "@/lib/types";
 import { useDownloads } from "@/hooks/use-downloads";
 import { useMountSubscription } from "@/hooks/use-mount-subscription";
-import { isManagedServeRuntimeTarget } from "@/lib/serve-runtime";
-import { describeFailedEngineJob } from "@/features/settings/runtime-targets";
-import { buildStarterRecipe } from "./setup-helpers";
 import {
-  CONTROLLER_UNREACHABLE_MESSAGE,
-  finishRuntimeJobEffect,
-  formatLoadWarning,
-  requestEffect,
-  setupErrorMessage,
-  withSetupTimeoutEffect,
-} from "./use-setup-effects";
+  loadSecondarySetupDataEffect,
+  loadSetupDataEffect,
+  refreshRuntimeStateEffect,
+  type SetupLoadSetters,
+} from "./setup-load";
+import {
+  beginDownloadEffect,
+  configureAndLaunchEffect,
+  connectRemotePresetEffect,
+  markSetupComplete,
+  runRuntimeJobEffect,
+  saveSettingsEffect,
+} from "./setup-actions";
 import { useSetupBenchmark } from "./use-setup-benchmark";
 
 type ManagedSetupBackend = Extract<EngineBackend, "vllm" | "sglang" | "mlx">;
-
-const markSetupComplete = (): void => {
-  try {
-    localStorage.setItem("local-studio-setup-complete", "true");
-  } catch {}
-};
 
 export function useSetup() {
   const router = useRouter();
@@ -81,120 +77,32 @@ export function useSetup() {
   }, [downloadsState.downloads, selectedModel]);
 
   const refreshRuntimeState = useCallback(() => {
-    return Effect.runPromise(
-      Effect.gen(function* () {
-        const [targetPayload, jobPayload] = yield* Effect.all([
-          requestEffect(() => api.getRuntimeTargets()).pipe(
-            Effect.catch(() => Effect.succeed({ targets: [] })),
-          ),
-          requestEffect(() => api.getRuntimeJobs()).pipe(
-            Effect.catch(() => Effect.succeed({ jobs: [] })),
-          ),
-        ] as const);
-        setRuntimeTargets(targetPayload.targets);
-        setRuntimeJobs(jobPayload.jobs);
-      }),
-    );
+    return Effect.runPromise(refreshRuntimeStateEffect({ setRuntimeTargets, setRuntimeJobs }));
   }, []);
 
-  const loadSecondarySetupData = useCallback((initialWarnings: string[]) => {
-    return Effect.runPromise(
-      Effect.gen(function* () {
-        const warnings = [...initialWarnings];
-        const [recommendationsResult, presetsResult, targetResult, jobResult] = yield* Effect.all([
-          Effect.result(
-            withSetupTimeoutEffect(api.getModelRecommendations(), "model recommendations"),
-          ),
-          Effect.result(withSetupTimeoutEffect(api.getStarterPresets(), "starter presets")),
-          Effect.result(withSetupTimeoutEffect(api.getRuntimeTargets(), "runtime targets")),
-          Effect.result(withSetupTimeoutEffect(api.getRuntimeJobs(), "runtime jobs")),
-        ] as const);
-
-        if (Result.isSuccess(presetsResult)) {
-          setPresets(presetsResult.success.presets || []);
-        } else setPresets([]);
-
-        if (Result.isSuccess(recommendationsResult)) {
-          setRecommendations(recommendationsResult.success.recommendations || []);
-          setMaxVram(recommendationsResult.success.max_vram_gb ?? 0);
-        } else {
-          setRecommendations([]);
-          setMaxVram(0);
-          warnings.push(
-            `model recommendations: ${setupErrorMessage(recommendationsResult.failure)}`,
-          );
-        }
-
-        if (Result.isSuccess(targetResult)) {
-          setRuntimeTargets(targetResult.success.targets);
-        } else {
-          setRuntimeTargets([]);
-          warnings.push(`runtime targets: ${setupErrorMessage(targetResult.failure)}`);
-        }
-
-        if (Result.isSuccess(jobResult)) {
-          setRuntimeJobs(jobResult.success.jobs);
-        } else {
-          setRuntimeJobs([]);
-          warnings.push(`runtime jobs: ${setupErrorMessage(jobResult.failure)}`);
-        }
-
-        setLoadWarning(formatLoadWarning(warnings));
-      }),
-    );
-  }, []);
+  const loadSecondarySetupData = useCallback(
+    (initialWarnings: string[], loadSetters: SetupLoadSetters) => {
+      return Effect.runPromise(loadSecondarySetupDataEffect(initialWarnings, loadSetters));
+    },
+    [],
+  );
 
   const loadSetupData = useCallback(() => {
+    const loadSetters: SetupLoadSetters = {
+      setLoading,
+      setError,
+      setLoadWarning,
+      setSettings,
+      setModelsDir,
+      setDiagnostics,
+      setRecommendations,
+      setMaxVram,
+      setRuntimeTargets,
+      setRuntimeJobs,
+      setPresets,
+    };
     return Effect.runPromise(
-      Effect.gen(function* () {
-        setLoading(true);
-        setError(null);
-        setLoadWarning(null);
-        const warnings: string[] = [];
-        const [settingsResult, diagnosticsResult] = yield* Effect.all([
-          Effect.result(withSetupTimeoutEffect(api.getStudioSettings(), "settings")),
-          Effect.result(
-            withSetupTimeoutEffect(api.getStudioDiagnostics(), "controller diagnostics"),
-          ),
-        ] as const);
-
-        if (Result.isSuccess(settingsResult)) {
-          setSettings(settingsResult.success);
-          setModelsDir(settingsResult.success.effective.models_dir);
-        } else {
-          setSettings(null);
-          warnings.push(`settings: ${setupErrorMessage(settingsResult.failure)}`);
-        }
-
-        if (Result.isSuccess(diagnosticsResult)) {
-          setDiagnostics(diagnosticsResult.success);
-          if (Result.isFailure(settingsResult)) {
-            setModelsDir(diagnosticsResult.success.config.models_dir || "");
-          }
-        } else {
-          setDiagnostics(null);
-          warnings.push(`controller diagnostics: ${setupErrorMessage(diagnosticsResult.failure)}`);
-        }
-
-        if (Result.isFailure(settingsResult) && Result.isFailure(diagnosticsResult)) {
-          setError(CONTROLLER_UNREACHABLE_MESSAGE);
-          return;
-        }
-
-        setRecommendations([]);
-        setMaxVram(0);
-        setRuntimeTargets([]);
-        setRuntimeJobs([]);
-        setLoadWarning(formatLoadWarning(warnings));
-
-        void loadSecondarySetupData(warnings);
-      }).pipe(
-        Effect.ensuring(
-          Effect.sync(() => {
-            setLoading(false);
-          }),
-        ),
-      ),
+      loadSetupDataEffect(loadSetters, (warnings) => loadSecondarySetupData(warnings, loadSetters)),
     );
   }, [loadSecondarySetupData]);
 
@@ -209,26 +117,14 @@ export function useSetup() {
     }
     setSavingSettings(true);
     return Effect.runPromise(
-      Effect.gen(function* () {
-        const result = yield* requestEffect(() =>
-          api.updateStudioSettings({ models_dir: modelsDir.trim() }),
-        );
-        setSettings(result);
-        setModelsDir(result.effective.models_dir);
-        setHardwareConfirmed(false);
-        setStep(1);
-      }).pipe(
-        Effect.catch((err) =>
-          Effect.sync(() =>
-            setError(err instanceof Error ? err.message : "Failed to update settings"),
-          ),
-        ),
-        Effect.ensuring(
-          Effect.sync(() => {
-            setSavingSettings(false);
-          }),
-        ),
-      ),
+      saveSettingsEffect(modelsDir, {
+        setSettings,
+        setModelsDir,
+        setHardwareConfirmed,
+        setStep,
+        setError,
+        setSavingSettings,
+      }),
     );
   }, [modelsDir]);
 
@@ -237,31 +133,13 @@ export function useSetup() {
       setUpgrading(true);
       setError(null);
       return Effect.runPromise(
-        Effect.gen(function* () {
-          const { job } = yield* requestEffect(() => api.createRuntimeJob(payload));
-          setRuntimeJobs((current) => [
-            job,
-            ...current.filter((candidate) => candidate.id !== job.id),
-          ]);
-          const finalJob = yield* finishRuntimeJobEffect(job.id, setRuntimeJobs);
-          if (finalJob.status === "error") {
-            setError(describeFailedEngineJob(finalJob));
-          }
-          const refreshed = yield* requestEffect(() => api.getStudioDiagnostics());
-          setDiagnostics(refreshed);
-        }).pipe(
-          Effect.catch((err) =>
-            Effect.sync(() => setError(err instanceof Error ? err.message : "Runtime job failed")),
-          ),
-          Effect.ensuring(
-            Effect.gen(function* () {
-              yield* requestEffect(() => refreshRuntimeState()).pipe(
-                Effect.catch(() => Effect.void),
-              );
-              setUpgrading(false);
-            }),
-          ),
-        ),
+        runRuntimeJobEffect(payload, {
+          setError,
+          setRuntimeJobs,
+          setDiagnostics,
+          setUpgrading,
+          refreshRuntimeState,
+        }),
         { signal: lifecycle.abort.signal },
       ).catch(() => undefined);
     },
@@ -292,19 +170,11 @@ export function useSetup() {
       setCreatedRecipeId(null);
       resetBenchmark();
       return Effect.runPromise(
-        requestEffect(() =>
-          downloadsState.startDownload({
-            model_id: modelId,
-            ...(preset?.allow_patterns?.length ? { allow_patterns: preset.allow_patterns } : {}),
-          }),
-        ).pipe(
-          Effect.map(() => setStep(3)),
-          Effect.catch((err) =>
-            Effect.sync(() =>
-              setError(err instanceof Error ? err.message : "Failed to start download"),
-            ),
-          ),
-        ),
+        beginDownloadEffect(modelId, preset, {
+          startDownload: downloadsState.startDownload,
+          setStep,
+          setError,
+        }),
       );
     },
     [downloadsState, resetBenchmark],
@@ -332,35 +202,11 @@ export function useSetup() {
       setConnectingRemote(true);
       setRemoteError(null);
       return Effect.runPromise(
-        Effect.gen(function* () {
-          const existing = yield* requestEffect(() => api.getProviders()).pipe(
-            Effect.catch(() => Effect.succeed({ providers: [] })),
-          );
-          const alreadyThere = existing.providers.some((provider) => provider.id === preset.id);
-          if (alreadyThere) {
-            yield* requestEffect(() =>
-              api.updateProvider(preset.id, { api_key: apiKey, enabled: true }),
-            );
-          } else {
-            yield* requestEffect(() =>
-              api.createProvider({
-                id: preset.id,
-                name: preset.name,
-                base_url: remote.base_url,
-                api_key: apiKey,
-              }),
-            );
-          }
-          markSetupComplete();
-          router.push("/agent?new=1");
-        }).pipe(
-          Effect.catch((err) =>
-            Effect.sync(() =>
-              setRemoteError(err instanceof Error ? err.message : "Failed to connect provider"),
-            ),
-          ),
-          Effect.ensuring(Effect.sync(() => setConnectingRemote(false))),
-        ),
+        connectRemotePresetEffect(preset, remote, apiKey, {
+          setRemoteError,
+          setConnectingRemote,
+          openAgentChat: () => router.push("/agent?new=1"),
+        }),
       );
     },
     [remoteApiKey, router],
@@ -386,58 +232,9 @@ export function useSetup() {
     resetBenchmark();
 
     return Effect.runPromise(
-      Effect.gen(function* () {
-        const backend = selectedPreset?.backend ?? "vllm";
-        const targetPayload = yield* requestEffect(() => api.getRuntimeTargets()).pipe(
-          Effect.catch(() => Effect.succeed({ targets: [] satisfies RuntimeTarget[] })),
-        );
-        const runtimeInstalled = targetPayload.targets.some((target) =>
-          backend === "llamacpp"
-            ? target.backend === backend && target.installed
-            : isManagedServeRuntimeTarget(backend, target) && target.installed,
-        );
-        if (!runtimeInstalled) {
-          const { job } = yield* requestEffect(() =>
-            api.createRuntimeJob({ backend, type: "install" }),
-          );
-          setRuntimeJobs((current) => [
-            job,
-            ...current.filter((candidate) => candidate.id !== job.id),
-          ]);
-          const finalJob = yield* finishRuntimeJobEffect(job.id, setRuntimeJobs);
-          if (finalJob.status === "error") {
-            return yield* Effect.fail(new Error(describeFailedEngineJob(finalJob)));
-          }
-        }
-
-        let recipeId = createdRecipeId;
-        if (!recipeId) {
-          const existing = yield* requestEffect(() => api.getRecipes()).pipe(
-            Effect.catch(() => Effect.succeed({ recipes: [] })),
-          );
-          const recipe = buildStarterRecipe(activeDownload, existing.recipes, selectedPreset);
-          yield* requestEffect(() => api.createRecipe(recipe));
-          recipeId = recipe.id;
-          setCreatedRecipeId(recipe.id);
-        }
-
-        yield* requestEffect(() => api.launch(recipeId));
-        const ready = yield* requestEffect(() => api.waitReady(300));
-        if (!ready.ready) {
-          return yield* Effect.fail(
-            new Error(ready.error || "The model did not become ready in time."),
-          );
-        }
-
-        markSetupComplete();
-        setStep(5);
-      }).pipe(
-        Effect.catch((err) =>
-          Effect.sync(() =>
-            setLaunchError(err instanceof Error ? err.message : "Failed to configure and launch"),
-          ),
-        ),
-        Effect.ensuring(Effect.sync(() => setConfiguringRecipe(false))),
+      configureAndLaunchEffect(
+        { activeDownload, selectedPreset, createdRecipeId },
+        { setRuntimeJobs, setCreatedRecipeId, setStep, setLaunchError, setConfiguringRecipe },
       ),
     );
   }, [activeDownload, createdRecipeId, resetBenchmark, selectedPreset, setRuntimeJobs]);
